criando liste de clientes

+++ src/main/java/br/com/claricejoias_ws/dto/CompraDetalheDTO.java
@@ -5,6 +5,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -15,9 +16,9 @@ public class CompraDetalheDTO {
 
     private Long id;
     private LocalDateTime data;
-    private Double total;
+    private BigDecimal total;
     private String metodoPagamento; // "especie", "pix", "cartao", "fiado"
-    private Double valorEntrada;    // Pode ser null ou 0 se não for fiado
+    private BigDecimal valorEntrada;    // Pode ser null ou 0 se não for fiado
     private Integer parcelas;       // 1 para a vista, >1 para parcelado
 
     // Se quiser retornar os itens comprados no futuro, pode adicionar:
