+++ src/main/java/br/com/claricejoias_ws/dto/PagamentoRequestDTO.java
@@ -2,10 +2,12 @@ package br.com.claricejoias_ws.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class PagamentoRequestDTO {
     private String metodo;
     private Integer parcelas;
-    private Double valorRecebido;
-    private Double valorEntrada; // <-- NOVO CAMPO ADICIONADO
+    private BigDecimal valorRecebido;
+    private BigDecimal valorEntrada; // <-- NOVO CAMPO ADICIONADO
 }
