+++ src/main/java/br/com/claricejoias_ws/dto/VendaRequestDTO.java
@@ -1,12 +1,14 @@
 package br.com.claricejoias_ws.dto;
 
 import lombok.Data;
+
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
 public class VendaRequestDTO {
     private List<ItemVendaRequestDTO> itens;
-    private Double total;
+    private BigDecimal total;
     private PagamentoRequestDTO pagamento;
     private ClienteRequestDTO cliente;
 }
